adding in testing before bothering to run benchmark
