+++ testing/conformance/benches/bencher.rs
@@ -9,18 +9,19 @@ use std::io::BufReader;
 use std::iter;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
-use log::*;
+
 use conformance_tests::test_utils::*;
 use conformance_tests::vector::{MessageVector, Selector, TestVector, Variant};
 use conformance_tests::vm::{TestKernel, TestMachine};
 use criterion::*;
 use fvm::executor::{ApplyKind, DefaultExecutor, Executor};
 use fvm_shared::address::Protocol;
+use fvm_shared::blockstore::MemoryBlockstore;
 use fvm_shared::crypto::signature::SECP_SIG_LEN;
 use fvm_shared::encoding::Cbor;
 use fvm_shared::message::Message;
+use log::*;
 use walkdir::WalkDir;
-use fvm_shared::blockstore::MemoryBlockstore;
 
 fn apply_messages(messages: &mut Vec<(Message, usize)>, exec: &mut DefaultExecutor<TestKernel>) {
     // Apply all messages in the vector.
@@ -31,18 +32,19 @@ fn apply_messages(messages: &mut Vec<(Message, usize)>, exec: &mut DefaultExecut
     }
 }
 
-fn bench_vector_variant(group: &mut BenchmarkGroup<measurement::WallTime>,
-                        name: String,
-                        variant: &Variant,
-                        v: &MessageVector,
-                        bs: &MemoryBlockstore,
+fn bench_vector_variant(
+    group: &mut BenchmarkGroup<measurement::WallTime>,
+    name: String,
+    variant: &Variant,
+    v: &MessageVector,
+    bs: &MemoryBlockstore,
 ) {
     group.bench_function(name, move |b| {
         b.iter_batched_ref(
             || {
                 let v = v.clone();
                 let bs = bs.clone();
-                // TODO next few lines don't impact the benchmarks, but it makes them run more slowly... ought to make a base copy of the machine and exec and deepcopy them each time.
+                // TODO next few lines don't impact the benchmarks, but it makes them run waaaay more slowly... ought to make a base copy of the machine and exec and deepcopy them each time.
                 let machine = TestMachine::new_for_vector(v, variant, bs);
                 machine.load_builtin_actors_modules().unwrap();
                 let exec: DefaultExecutor<TestKernel> = DefaultExecutor::new(machine);
@@ -75,15 +77,21 @@ fn bench_vector_file(group: &mut BenchmarkGroup<measurement::WallTime>, path: Pa
     let TestVector::Message(vector) = vector;
     let skip = !vector.selector.as_ref().map_or(true, Selector::supported);
     if skip {
-        info!("skipped benching {}, selector not supported.", path.display());
+        info!(
+            "skipped benching {}, selector not supported.",
+            path.display()
+        );
         return;
     }
 
-    // TODO ought to run the relevant conformance test and ensure it passes before we run the benchmark.
     let (bs, _) = async_std::task::block_on(vector.seed_blockstore()).unwrap();
 
     for variant in vector.preconditions.variants.iter() {
         let name = format!("{} | {}", path.display(), variant.id);
+        // TODO correctly handle a failing test please.
+        // this tests the variant before we run it and record the results back to criterion.
+        // if we broke the test, it's not a valid optimization :P
+        run_variant(bs.clone(), &vector, variant).unwrap();
         bench_vector_variant(group, name, variant, &vector, &bs);
     }
 }
@@ -103,14 +111,18 @@ fn bench(c: &mut Criterion) {
     // pretty_env_logger::init();
 
     // TODO match globs?
-    let (mut vector_results, _is_many) : (Vec<PathBuf>, bool) = match var("VECTOR") {
-        Ok(v) => (iter::once(Path::new(v.as_str()).to_path_buf()).collect(), false),
+    let (mut vector_results, _is_many): (Vec<PathBuf>, bool) = match var("VECTOR") {
+        Ok(v) => (
+            iter::once(Path::new(v.as_str()).to_path_buf()).collect(),
+            false,
+        ),
         Err(_) => (
             WalkDir::new("test-vectors/corpus")
                 .into_iter()
                 .filter_map(|e| e.ok())
                 .filter(is_runnable)
-                .map(|e| e.path().to_path_buf()).collect(),
+                .map(|e| e.path().to_path_buf())
+                .collect(),
             true,
         ),
     };
